requirements.txt
amqp==5.2.0
anyio==4.2.0
asgiref==3.7.2
billiard==4.2.0
boto3==1.34.43
botocore==1.34.43
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.9
django-appconf==1.0.6
django-cors-headers==4.3.1
django-ipware==6.0.4
django-storages==1.14.2
djangorestframework==3.14.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
jmespath==1.0.1
kombu==5.3.5
oauthlib==3.2.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ipware==2.0.1
python-telegram-bot==20.7
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
